fibonacci numbers sequence

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,17 +5,17 @@
  */
 int main(void)
 {
-	unsigned long long int c, a = 1, b = 2;
+	unsigned long int c, a = 1, b = 2;
 	int i;
 
-	printf("%llu, %llu, ", a, b);
+	printf("%ld, %ld, ", a, b);
 	for (i = 1; i <= 48; i++)
 	{
 		c = a + b;
 		if (i != 48)
-			printf("%llu, ", c);
+			printf("%ld, ", c);
 		else
-			printf("%llu", c);
+			printf("%ld", c);
 		a = b;
 		b = c;
 	}
